cricket project based on python
One of the most treasured memories of growing up is playing cricket with friends. Just like me, there are countless other people whose favourite part of the day is to grab a pair of bat and ball and just play. Gully cricket is the most popular form of sport, is played in everywhere. So, today’s project is going to be very helpful for those cricket fans who want to create their own decision review system.